flask_app/order/requirements.txt
# REST API Generation
flask==1.1.1

# DB & ORM (Object Relational Mapping)
sqlalchemy==1.3.8
mysql-connector-python==8.0.17

# To load .env variables
python-dotenv==0.10.3

# For error handling:
werkzeug==0.15.6

# As a production server
gunicorn==19.9.0

requests==2.26.0

#Pika
token==1.2.0
pika

# PyJWT
PyJWT

# bcrypt
bcrypt

# crypto
pycryptodome
cryptography

#test
pytest==5.3.1
pytest-cov==2.8.1